+++ packages/payment-integration-api/src/index.ts
@@ -58,6 +58,7 @@ export {
     RequestError,
     StandardError,
     StorefrontErrorResponseBody,
+    TimeoutError,
     isRequestError,
 } from './errors';
 export {