+++ packages/paypal-commerce-integration/src/paypal-commerce-credit/paypal-commerce-credit-button-strategy.spec.ts
@@ -1,6 +1,3 @@
-import { createFormPoster, FormPoster } from '@bigcommerce/form-poster';
-import { createRequestSender, RequestSender } from '@bigcommerce/request-sender';
-import { getScriptLoader } from '@bigcommerce/script-loader';
 import { EventEmitter } from 'events';
 
 import {
@@ -19,14 +16,15 @@ import {
     PaymentIntegrationServiceMock,
 } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
 
-import getBillingAddressFromOrderDetails from '../mocks/get-billing-address-from-order-details.mock';
-import getPayPalCommerceOrderDetails from '../mocks/get-paypal-commerce-order-details.mock';
-import getShippingAddressFromOrderDetails from '../mocks/get-shipping-address-from-order-details.mock';
-import { getPayPalCommercePaymentMethod } from '../mocks/paypal-commerce-payment-method.mock';
-import { getPayPalSDKMock } from '../mocks/paypal-sdk.mock';
+import {
+    getBillingAddressFromOrderDetails,
+    getPayPalCommerceIntegrationServiceMock,
+    getPayPalCommerceOrderDetails,
+    getPayPalCommercePaymentMethod,
+    getPayPalSDKMock,
+    getShippingAddressFromOrderDetails,
+} from '../mocks';
 import PayPalCommerceIntegrationService from '../paypal-commerce-integration-service';
-import PayPalCommerceRequestSender from '../paypal-commerce-request-sender';
-import PayPalCommerceScriptLoader from '../paypal-commerce-script-loader';
 import {
     PayPalCommerceButtonsOptions,
     PayPalCommerceHostWindow,
@@ -40,16 +38,12 @@ describe('PayPalCommerceCreditButtonStrategy', () => {
     let buyNowCart: Cart;
     let cart: Cart;
     let eventEmitter: EventEmitter;
-    let formPoster: FormPoster;
-    let requestSender: RequestSender;
     let strategy: PayPalCommerceCreditButtonStrategy;
     let paymentIntegrationService: PaymentIntegrationService;
     let paymentMethod: PaymentMethod;
     let paypalButtonElement: HTMLDivElement;
     let paypalMessageElement: HTMLDivElement;
     let paypalCommerceIntegrationService: PayPalCommerceIntegrationService;
-    let paypalCommerceRequestSender: PayPalCommerceRequestSender;
-    let paypalCommerceScriptLoader: PayPalCommerceScriptLoader;
     let paypalSdk: PayPalSDK;
 
     const defaultMethodId = 'paypalcommercecredit';
@@ -115,21 +109,10 @@ describe('PayPalCommerceCreditButtonStrategy', () => {
 
         eventEmitter = new EventEmitter();
 
+        paypalCommerceIntegrationService = getPayPalCommerceIntegrationServiceMock();
         paymentMethod = getPayPalCommercePaymentMethod();
         paypalSdk = getPayPalSDKMock();
-
-        formPoster = createFormPoster();
-        requestSender = createRequestSender();
         paymentIntegrationService = <PaymentIntegrationService>new PaymentIntegrationServiceMock();
-        paypalCommerceRequestSender = new PayPalCommerceRequestSender(requestSender);
-        paypalCommerceScriptLoader = new PayPalCommerceScriptLoader(getScriptLoader());
-
-        paypalCommerceIntegrationService = new PayPalCommerceIntegrationService(
-            formPoster,
-            paymentIntegrationService,
-            paypalCommerceRequestSender,
-            paypalCommerceScriptLoader,
-        );
 
         strategy = new PayPalCommerceCreditButtonStrategy(
             paymentIntegrationService,